+++ packages/ui-react/src/index.ts
@@ -1,5 +1,12 @@
 export { cn } from './lib/utils';
 export * from './components/ui/breadcrumb';
 export * from './components/ui/button';
+export * from './components/ui/checkbox';
+export * from './components/ui/radio';
+export * from './components/ui/input';
+export * from './components/ui/search';
+export * from './components/ui/select';
 export * from './components/ui/button-icon';
 export * from './components/ui/switch';
+export * from './components/ui/tooltip';
+export * from './components/ui/tag';